src/requirements.txt
# API & Web
fastapi==0.118.0
uvicorn[standard]==0.37.0
python-multipart==0.0.9
python-dotenv==1.0.1
pydantic-settings==2.2.1
aiofiles==23.2.1

# LLMs / Vector / PDFs
langchain==0.3.27
PyMuPDF==1.26.4
openai==2.1.0
cohere==5.16.1
qdrant-client==1.12.2

# Databases
SQLAlchemy==2.0.43
asyncpg==0.30.0
alembic==1.16.5
# Switched from psycopg2 -> psycopg (v3) binary wheels for 3.12/3.13
psycopg[binary]>=3.1,<3.4
pgvector==0.4.0

# Mongo (see note below about deprecation)
motor==3.5.0
pymongo>=4.0,<5.0
pydantic-mongo==2.3.0

# NLP
nltk==3.9.2

# Monitoring and metrics
prometheus-client==0.23.1
starlette-exporter==0.23.0
fastapi-health==0.4.0

# Task Queue and Background Processing
celery==5.5.3
# Kombu/Celery currently cap redis versions; keep within supported range
redis>=4.5.2,!=5.0.2,<=5.2.1
kombu==5.5.4
billiard==4.2.2
vine==5.1.0
flower==2.0.1
